+++ pkg/pipeline/pipeline.go
@@ -410,6 +410,13 @@ type EntityAttribute struct {
 	Attribute string `json:"attribute"`
 }
 
+type UpstreamColumn struct {
+	Asset      string `json:"asset" yaml:"asset,omitempty" mapstructure:"asset"`
+	Column     string `json:"column" yaml:"column,omitempty" mapstructure:"column"`
+	Table      string `json:"table" yaml:"table,omitempty" mapstructure:"table"`
+	AssetFound bool   `json:"asset_found" yaml:"asset_found,omitempty" mapstructure:"asset_found"`
+}
+
 type Column struct {
 	EntityAttribute *EntityAttribute `json:"entity_attribute" yaml:"-" mapstructure:"-"`
 	Name            string           `json:"name" yaml:"name,omitempty" mapstructure:"name"`
@@ -419,6 +426,7 @@ type Column struct {
 	UpdateOnMerge   bool             `json:"update_on_merge" yaml:"update_on_merge,omitempty" mapstructure:"update_on_merge"`
 	Extends         string           `json:"-" yaml:"extends,omitempty" mapstructure:"extends"`
 	Checks          []ColumnCheck    `json:"checks" yaml:"checks,omitempty" mapstructure:"checks"`
+	Upstreams       []UpstreamColumn `json:"upstreams" yaml:"upstreams,omitempty" mapstructure:"upstreams"`
 }
 
 func (c *Column) HasCheck(check string) bool {
@@ -758,7 +766,7 @@ func (a *Asset) EnrichFromEntityAttributes(entities []*glossary.Entity) error {
 }
 
 func (a *Asset) PropagateColumns(foundPipeline *Pipeline) error {
-	return parseLineageRecursive(foundPipeline, a)
+	return (&LineageExtractor{}).ColumnLineage(foundPipeline, a)
 }
 
 func (a *Asset) Persist(fs afero.Fs) error {
@@ -924,7 +932,7 @@ type Pipeline struct {
 
 func (p *Pipeline) PropagateColumns() error {
 	for _, asset := range p.Assets {
-		if err := parseLineageRecursive(p, asset); err != nil {
+		if err := (&LineageExtractor{}).ColumnLineage(p, asset); err != nil {
 			return err
 		}
 	}
